refactor _create_bare_model to shared method
